+++ src/vue/modules/requests/update-asset-requests/wrappers/update-asset-request.js
@@ -33,4 +33,12 @@ export class UpdateAssetRequest extends Request {
   get isWithdrawable () {
     return Boolean(this.policy & ASSET_POLICIES.withdrawable)
   }
+
+  get isBaseInAtomicSwap () {
+    return Boolean(this.policy & ASSET_POLICIES.canBeBaseInAtomicSwap)
+  }
+
+  get isQuoteInAtomicSwap () {
+    return Boolean(this.policy & ASSET_POLICIES.canBeQuoteInAtomicSwap)
+  }
 }