+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Added
+- New Atomic swap feature:
+  - Create atomic swap form
+- `canBeBaseInAtomicSwap` and `canBeQuoteInAtomicSwap` policy for assets
+
 ## [1.9.0-rc.0] - 2019-06-24
 #### Added
 - New "My assets" page